+++ docs/versions.js
@@ -1,0 +1,8 @@
+var DOC_VERSIONS = [
+    "stable",
+    "v0.3",
+    "v0.2",
+    "v0.1",
+    "dev",
+  ];
+