+++ docs/src/changelog.md
@@ -1,1 +1,6 @@
 ## Unreleased
+
+
+## v0.1.3
+- Added interpolation for GeoArrays
+- Added FOSS4G notebook